Split code over multiple header and cpp files
In practice, it is best to split up the code into smaller, more
manageable chunks. This works best when distinct aspects of the
functionality can be grouped into logical 'units' or modules, which can
be inspected and understood independently.

To do this, we need to place the *declarations* of our various functions
in the *header* file, so that other parts of the code have the
information required to use them (name, number and type of arguments,
return type, etc). We then place the *definitions* in a separate cpp
file (typically with the same name), which can be *compiled* separately,
and the resulting object file can then be *linked* into the final
executable at a later stage.

This code can be compiled and linked using the following commands:

g++ -std=c++20 -c shotgun.cpp -o shotgun.o
g++ -std=c++20 -c fragments.cpp -o fragments.o
g++ shotgun.o fragments.o -o shotgun