Add -v command-line option for more verbose output
Most practical commands accept *options* in addition to their mandatory
arguments. Many commands accept an option to enable some form of
*verbose mode*, which provides more detailed feedback information about
the process, and is very useful when debugging.

Here, we look for the "-v" option in the argument list, and if found, we
enable verbose mode and remove the option from the argument list.  We
also store the information about whether verbose mode is enabled in a
*global variable*. Global variables are almost always a bad idea, but
are acceptable when used to represent an entity that is guaranteed to be
unique (std::cout / std::cerr are good examples of this). However, to
avoid naming conflicts and reduce the scope for name shadowing, it is
best to declare this variable within its own bespoke *namespace*.