Add function to find biggest overlap
This time, we need to return multiple pieces of information from the
call: the overlap, and the index of the overlapping fragment. There are
several ways to do this, but probably the clearest is to define a new
datatype, a *struct*, to hold the variables we need together in a single
object, which we can then return as a single entity. To ensure the code
remains clear and easy to follow, make sure to pick sensible names for
the member variables!