Avoiding needless copies: using std::string_view
The main computational overhead in our program is the formation and
comparison of the substrings corresponding to the overlapping regions.
When extracting the substrings into std::string variables, this
necessarily involves a *deep copy*: the characters have to be copied
over into the new std::string, and this involves memory allocations &
deallocations, along with the actual copying operations.

However, we can avoid the need to make these copies using the
std::string_view class, which provides a non-owning *view* of the
string, with its own offset and length. We can use this to specify the
substrings we want to compare, without making actual standalone copies
of them.

You can verify the improved runtime using the 'time' command.