Create the ShotgunSequencer class to represent the shotgun sequencing algorithm
In C++, *classes* is how we define new objects. A class extends the
concept of a struct by adding *methods* to imbue behaviour into the
object. Methods provide ways to interact with the object, without
exposing the data within the object directly.

Note that structs can have methods too, and are essentially equivalent
to classes, the main difference is that members are public by default in
structs, and private by default in classes. In practice, classes are
almost invariably preferred, unless the object is only destined to be
used as a lightweight container for a simple collection of data.