Create Image class
We create a simple Image class to hold the image data and associated
information. To ensure encapsulation, we make the member variables
private, and provide read-only accessor methods. The constructor is
responsible for initialising the class from the information provided,
and making sure the data are all self-consistent.

We also introduce the std::array template class, which is similar to
std::vector, but is used to hold a fixed-size array of elements. This is
more efficient when the size of the array is a known constant. Here, we
use it to hold the dimensions of the image, and since this is a 2D
image, the size is known to be 2.