Add Dataset class
We add another class to hold the full dataset, which simply abstracts
the std::vector<Image>, providing the requisite functions for loading
and accessing the data.

We pass the list of files as a std::vector<std::string>, which we can
construct from argv & argc using one of the std::vector constructors.