Introduce plotting on the terminal
We have provided a small library that you can use to plot on the
terminal. This uses an ancient technology called sixel graphics
(originally designed for dot matrix printers!), which can still be used
with the right terminal. The terminal provided by VSCode (as part of
GitHub codespaces) happens to support this straight away (you may have
to 'enable images in the terminal' in the settings for VSCode for this
to work).