Introduce template programming for Image class
This allows us to define a class template, capable of representing an
image where each pixel value is represented using any type of our
choosing. Here, we still use an Image of int in the program, but our
Image class is now capable of representing data of other types if we
need to.

This is particularly useful to store the image of correlation
coefficients, which are best stored as floating-point values in the
range [ -1, 1 ].