Convert the Dataset into a template class as well
We can also use template programming for the Dataset class, making it
capable of holding a set of images of arbitrary types for the pixel
values.

In this case however, we need to make sure any template function
or method definitions are also included in the header. If left in the
cpp file, the compiler will not be able to generate the code for the
concrete realisation of the class where it is needed. For example, if we
use the Dataset<int> class in our main run() function, the compiler will
need to have access to the code for all the functions and methods we
rely on when it compiles the fmri.cpp file (where the run() function is
defined).

Conversely, if we left the definition of these functions in
the dataset.cpp file, the compiler wouldn't know which versions we might
be using in any other parts of our program when compiling dataset.cpp.
Should it generate code for Dataset<int>, Dataset<float>,
Dataset<double>, etc.? In general, there are too many possible types
here, so the compiler will only check for correct syntax: it will
generate no code at all. This will lead to link-time errors later on,
which look like 'undefined symbol' errors.