Final touches to project
We finalise the project by removing functionality that is no longer
being used, and change the behaviour of the program slightly: by
default, the map of correlation coefficients is computed and displayed.
However, if the -p x y option is provided, the timecourse for that pixel
is shown along with the task, and only the correlation coefficient for
this timecourse is computed and displayed.