Define the Segment base class
We set up the bare-bones outline of what our base class might be for one
of the segments of the robot arm. At the very least, it needs a
constructor, which will also need to initialise a reference to the next
segment on the arm.  We also add a private member denoting the type of
the segment.

In anticipation of adding more types of segments later, we set up a
dedicated "Segment" namespace to hold them. Finally, to ensure our code
remains organised, we create a dedicated "segment" folder to place the
related code.