Compute and display speed of the tip point
While this could be done using simpler methods, here we introduce an
operator overload for the subtraction operator between Points, and a
small utility function to compute the length of a Point vector.
Together, these allow us to compute the speed of the tip in a readable
manner.

We also define a function to compute and display the trajectory and
corresponding speed on the terminal.