Add explicit constructor with default argument
We add an explicit constructor to allow the user to set the minimum
overlap in our ShotgunSequencer algorithm. This takes a single integer
argument, whose value defaults to 10. This in effect provides us with a
default constructor (when invoked with no argument), while allowing us
to specify the minimum overlap if the default is not appropriate.

Note that we need to initialise the minimum overlap in the constructor
as we have declared it const.