Introduce template function for rescale() and correlation_coefficient()
C++ allows us to define template functions, that will operate on generic
types. The actual type to be used in the function is not know at the
point where it declared & defined. At the point where the function is
used, the compiler will substitute the desired type, and produce the
machine code for that specific version of the function at that point.

This allows us to define generic functions that can be used with any
(compatible) type, and avoid unnecessary code duplication.