Allow segments to be created by themselves
We modify the code to allow segments to be instantiated as standalone
objects, by setting the pointer to next to nullptr in the constructor.
We also provide a new method in the base class to allow each segment set
its next segment. This allows us to specify and modify how the robot arm is
configured at runtime.

To make it easier to specify the configuration, the new `set_next()`
method returns a reference to the next segment (i.e. the segment we
passed as the argument). This allows us to daisy-chain set_next() calls
with very little code.