Introduce AngleBase class with recursive set_angle() method
We introduce a method to set the angle using a recursive approach: we
specify the index of the angle, and the value to set. Each segment will
simply pass the call onto the next segment - unless that segment manages
an angle, in which case if will reduce the index value and pass it onto
the next segment, or if the index value is already zero, set its own
angle accordingly. This allows us to set all the angles without
reference to the specific instance for each of the relevant segments.

To simplify the implementation, we introduce another base class, derived
from the main base class, to represent segments that have an angle.
This class overrides the default set_angle() method provided by the main
base class (which simply forwards the call to the next segment), and
instead handles the angle as mentioned above (set its own angle if the
index is zero, or forward the call to the next segment with the index
decremented by one).