Use a smart pointer to manage the memory
We can use a smart pointer to manage all this for us. The
std::unique_ptr pointer is suitable here, as it ensures only one copy of
the pointed-to instance exists, and ensures it is deleted when the
unique_ptr is destroyed. This means we don't have to worry about the
destructor, copy constructor or assignment operator, and just leave
these as default values.

We can then replace the use of the new operator with the
std::make_unique() template call, which will create a unique pointer of
the requested type, and forward the arguments provided to the class
constructor.

Note that in the set_next() method, we need to use the swap() method to
transfer ownership of the unique_ptr from the argument to the data
member in the class. This means that after the method is called, the
original unique_ptr will no longer contain the segment is was originally
pointing to, which is fine as it is no longer required anyway.