Use new / delete to allow each segment to manage the next
We can implement manual memory management, by allowing each segment to
delete the next segment. We can set the next segment by providing a
pointer to a new instance of that segment, allocated on the heap using
new. The destructor for the base class can be delete the next instance.

Note that we rely on the fact that calling delete on a nullptr has no
effect and is safe (this is stated in the C++ standard).

We also need to make sure to delete any segment we may have currently
been managing when setting up the next one. The m_next pointer is
initialised as nullptr, and in most cases this statement will have no
effect - but in due course, we may have to modify the configuration at
runtime, in which case we need to ensure there are no memory leaks at
this point.

Finally, when setting up a non-default destructor, we will typically
also need to worry about the copy constructor, and the assignment
operator. Here, we opt to simply delete them, which means it will not be
possible to copy the Segments. This is one way to ensure there are no
double deletions, which would be concern if we duplicate the class
as-is, as the destructors of both copies will try to delete the same
address!

An alternative approach would be to perform deep copies in the copy
constructor / assignment operator, but this is unlikely to be necessary
and/or expected.